[GOBBLIN-248] Add javadocs and Rename DataType to InputType
